hooks: Teach sub-hooks to work on merges too
Add placeholder in merge commits too so the sub-hooks relying on
placeholder work.

With

    GIT_MERGE_AUTOEDIT=no

Git does not put default text message or any message for a clean
merge which can be used as a hook to insert placeholder so
just append placeholder at end in prepare-commit-msg sub-hook. This
does not affect a conflict merge, as eventually

    $ git commit

is required which has the default commit text which is used to insert
placeholder. With

    GIT_MERGE_AUTOEDIT=no

the merge commit message is taken as is so with

    $ git commit --amend --no-edit

the comments are included in the message. Without '--no-edit' option,
it's not an issue. To cover this case, strip comments in a commit-msg
sub-hook.

This also means the delete default text in
'53-massage-merge-commit-message' is technically not required but
better to leave it as is.