hooks: Fix '13-add-placeholder' sub-hook
There are cases when only the 'prepare-commit-msg' hook runs
(cherry-pick, amend with no edit, some rebase cases) which results in
the comments inserted by the 'prepare-commit-msg' sub-hook in the
commit message as the 'commit-msg' hook, thereby '97-remove-comments'
sub-hook, is not run. '97-remove-comments' sub-hook also does not work
with 'git commit -v' where it would remove the "scissors" resulting
in including the diff portion which is not prefixed with '#'.

 - With

       GIT_MERGE_AUTOEDIT=no

   add the placeholder for merges only in '13-add-placeholder'
   sub-hook and strip comments with '97-remove-comments' sub-hook,
   through 'commit-msg' hook, for merges only.
 - With

       GIT_MERGE_AUTOEDIT=yes

   add the placeholder using the default merge message in
   '13-add-placeholder' sub-hook.
 - '83-apply-commit-hook' sub-hook did not pass the "$2" arg to
   'commit-msg' hook and so '97-remove-comments' sub-hook did not get
   it. Teach it to do so.