find-merge: Optimize if commit is a first-parent
The logic to find if commit is a first-parent works but is
inefficient. The reason is that

    $ git rev-list --first-parent $commit..$branch

excludes $commit.

Printing the parents instead and selecting first-parent as

    $ git log --first-parent --pretty='%P' $commit..$branch |
        cut -d' ' -f1

includes $commit. This will not print the $branch commit but the case
where $commit is same as $branch is handled right before this.

On git's repo of ~50k commits, the time saved is 150ms-200ms.