hooks: Teach to also run local sub-hooks
Currently, sub-hooks are global and are run from '~/.git-hooks/<hook>.d'
dir. To run a local sub-hook for a repo, a global gitignored sub-hook
needs to be added which in turn calls the local sub-hook as,

    # path: ~/.git-hooks/commit-msg.d/99-local-sub-hook-(gitignore)
    if [[ -f .git/hooks/local-sub-hook ]]; then
        .git/hooks/local-sub-hook "$@"
    fi

So teach to also run local sub-hooks (if present) at
'.git/hooks/<hook>.d' dir while following the sort order of global and
local sub-hooks' filenames.