hooks: Remove 83-apply-commit-hook sub-hook
The commit-msg hook was not called for clean merge commit. So
'83-apply-commit-hook' was added to do that. This caused the editor
to show twice for merge commits. So 'GIT_MERGE_AUTOEDIT=no' was set
in .bashrc as workaround. This created edge cases as a side-effect.

The commit-msg hook is now called for clean merge commits in commit
f8b863598 ("builtin/merge: honor commit-msg hook for merges",
2017-09-07).

So remove workarounds.